[MOD] : chat activity + menu select fragment -> menu activity
