+++ android/app/src/main/java/com/kotlin/kiumee/presentation/menu/Menu.kt
@@ -1,0 +1,6 @@
+package com.kotlin.kiumee.presentation.menu
+
+data class Menu(
+    val name: String,
+    val price: String
+)
