+++ android/app/src/main/java/com/kotlin/kiumee/presentation/menu/MenuActivity.kt
@@ -1,13 +1,17 @@
-package com.kotlin.kiumee.presentation.chat
+package com.kotlin.kiumee.presentation.menu
 
+import android.content.Intent
 import androidx.recyclerview.widget.LinearSmoothScroller
 import androidx.recyclerview.widget.RecyclerView
 import com.kotlin.kiumee.R
 import com.kotlin.kiumee.core.base.BindingActivity
-import com.kotlin.kiumee.databinding.ActivityChatBinding
-import com.kotlin.kiumee.presentation.menuselect.MenuSelectFragment
+import com.kotlin.kiumee.databinding.ActivityMenuBinding
+import com.kotlin.kiumee.presentation.menu.chat.Chat
+import com.kotlin.kiumee.presentation.menu.chat.ChatAdapter
+import com.kotlin.kiumee.presentation.menu.chat.ChatItemDecoration
+import com.kotlin.kiumee.presentation.orderfinish.OrderFinishActivity
 
-class ChatActivity : BindingActivity<ActivityChatBinding>(R.layout.activity_chat) {
+class MenuActivity : BindingActivity<ActivityMenuBinding>(R.layout.activity_menu) {
     private val smoothScroller: RecyclerView.SmoothScroller by lazy {
         object : LinearSmoothScroller(this) {
             override fun getVerticalSnapPreference() = SNAP_TO_START
@@ -16,7 +20,6 @@ class ChatActivity : BindingActivity<ActivityChatBinding>(R.layout.activity_chat
 
     override fun initView() {
         initChatAdapter()
-        initActivityFragment()
 
         binding.btn.setOnClickListener {
             binding.rvChat?.layoutManager?.startSmoothScroll(
@@ -24,6 +27,35 @@ class ChatActivity : BindingActivity<ActivityChatBinding>(R.layout.activity_chat
                     targetPosition = 2
                 }
             )
+        }
+
+        // initTabPager()
+        initOrderBtnClickListener()
+    }
+
+    //    private fun initTabPager() {
+//        val tabLayout = binding.layoutFragmentMenuSelectTab
+//        val viewPager = binding.vpMenuSelect
+//
+//        val menuTabAdapter = MenuSelectAdapter(this)
+//        viewPager.adapter = menuTabAdapter
+//
+//        val tabs = listOf(
+//            "커피(HOT)",
+//            "커피(ICE)",
+//            "에이드&티",
+//            "음료",
+//            "디저트"
+//        )
+//
+//        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
+//            tab.text = tabs[position]
+//        }.attach()
+//    }
+
+    private fun initOrderBtnClickListener() {
+        binding.btnMenuCartOrder.setOnClickListener {
+            startActivity(Intent(this, OrderFinishActivity::class.java))
         }
     }
 
@@ -42,12 +74,5 @@ class ChatActivity : BindingActivity<ActivityChatBinding>(R.layout.activity_chat
         }
 
         binding.rvChat.addItemDecoration(ChatItemDecoration(this))
-    }
-
-    private fun initActivityFragment() {
-        supportFragmentManager.beginTransaction().apply {
-            replace(R.id.fcv_chat, MenuSelectFragment())
-            commit()
-        }
     }
 }
