+++ android/app/src/main/java/com/kotlin/kiumee/presentation/menu/MenuViewHolder.kt
@@ -1,0 +1,13 @@
+package com.kotlin.kiumee.presentation.menu
+
+import androidx.recyclerview.widget.RecyclerView
+import com.kotlin.kiumee.databinding.ItemMenuBinding
+
+class MenuViewHolder(private val binding: ItemMenuBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+    fun bind(data: Menu) {
+        when (binding) {
+            // dfdfa
+        }
+    }
+}
