--- android/app/src/main/java/com/kotlin/kiumee/presentation/menuselect/MenuSelect.kt
@@ -1,6 +1,0 @@
-package com.kotlin.kiumee.presentation.menuselect
-
-data class MenuSelect(
-    val name: String,
-    val price: String
-)
