--- android/app/src/main/java/com/kotlin/kiumee/presentation/menuselect/MenuSelectViewHolder.kt
@@ -1,13 +1,0 @@
-package com.kotlin.kiumee.presentation.menuselect
-
-import androidx.recyclerview.widget.RecyclerView
-import com.kotlin.kiumee.databinding.ItemMenuSelectBinding
-
-class MenuSelectViewHolder(private val binding: ItemMenuSelectBinding) :
-    RecyclerView.ViewHolder(binding.root) {
-    fun bind(data: MenuSelect) {
-        when (binding) {
-            // dfdfa
-        }
-    }
-}
