+++ android/app/src/main/java/com/kotlin/kiumee/presentation/orderfinish/OrderFinishActivity.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.lifecycleScope
 import com.kotlin.kiumee.R
 import com.kotlin.kiumee.core.base.BindingActivity
 import com.kotlin.kiumee.databinding.ActivityOrderFinishBinding
-import com.kotlin.kiumee.presentation.chat.ChatActivity
+import com.kotlin.kiumee.presentation.menu.MenuActivity
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
@@ -18,7 +18,7 @@ class OrderFinishActivity :
     private fun initTimer() {
         lifecycleScope.launch {
             delay(5000) // 5초
-            startActivity(Intent(this@OrderFinishActivity, ChatActivity::class.java))
+            startActivity(Intent(this@OrderFinishActivity, MenuActivity::class.java))
             // 현재 액티비티 닫기
             finish()
         }
